pyproject.toml
[tool.poetry]
name = "edupage_lunch"
version = "0.1.0"
description = "Script to automate ordering lunches on edupage"
authors = ["michqo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^23.3.0"
edupage-api = {git = "https://github.com/ivanhrabcak/edupage-api.git"}
orjson = "^3.9.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
